+++ src/main/java/aoc/y2024/day17/Part1.java
@@ -10,7 +10,7 @@ public class Part1 implements AocProblem<String> {
     @Override
     public String solve(List<String> lines) {
         var config = new Parser().parse(lines);
-        var output = new ArrayList<Long>();
+        var output = new ArrayList<Integer>();
         var computer = new Computer((value) -> output.add(value));
 
         computer.init(config.registerA(), config.registerB(), config.registerC());
